+++ react/__tests__/withCssHandles.test.tsx
@@ -1,31 +1,13 @@
-import React, { Component } from 'react'
-import withCssHandles from '../withCssHandles'
-import { useExtension } from '../hooks/useExtension'
+import React, { FC } from 'react'
 import { render } from '@vtex/test-tools/react'
 
-type ValueOf<T extends readonly any[]> = T[number]
-
-type CssHandlesInput = readonly string[]
-type CssHandles<T extends CssHandlesInput> = Record<ValueOf<T>, string>
-interface CssHandlesOptions {
-  migrationFrom?: string | string[]
-}
-
-interface Props<T extends CssHandlesInput> {
-  cssHandles?: CssHandles<T>
-}
-
-class ExampleComponent<T extends CssHandlesInput> extends Component<Props<T>> {
-  public render() {
-    const cssHandles = this.props.cssHandles as CssHandles<T>
-
-    return (
-      <div
-        data-testid="test-div"
-        className={Object.values(cssHandles).join(' ')}></div>
-    )
-  }
-}
+import withCssHandles from '../withCssHandles'
+import { useExtension } from '../hooks/useExtension'
+import type {
+  CssHandlesList,
+  CssHandlesBag,
+  CssHandlesOptions,
+} from '../CssHandlesTypes'
 
 jest.mock('../hooks/useExtension', () => ({
   useExtension: jest.fn(() => ({
@@ -38,13 +20,24 @@ jest.mock('../hooks/useExtension', () => ({
 
 describe('withCssHandles', () => {
   const renderComponent = (
-    handlesNames: CssHandlesInput,
-    options?: CssHandlesOptions
+    handlesNames: CssHandlesList,
+    options?: CssHandlesOptions<any>
   ) => {
-    const EnhancedComponent = withCssHandles<
-      typeof handlesNames,
-      Props<typeof handlesNames>
-    >(handlesNames, options)(ExampleComponent)
+    type InternalProps = CssHandlesBag<typeof handlesNames>
+
+    const ExampleComponent: FC<InternalProps> = ({ handles }) => {
+      return (
+        <div
+          data-testid="test-div"
+          className={Object.values(handles).join(' ')}
+        />
+      )
+    }
+
+    const EnhancedComponent = withCssHandles<typeof handlesNames>(
+      handlesNames,
+      options
+    )(ExampleComponent)
 
     return render(<EnhancedComponent />)
   }
@@ -52,22 +45,27 @@ describe('withCssHandles', () => {
   it('should apply proper classes to proper handles', () => {
     const CSS_HANDLES = ['element1', 'element2'] as const
     const { getByTestId } = renderComponent(CSS_HANDLES)
+
     expect(getByTestId('test-div').className).toBe(
       'vtex-app-2-x-element1 vtex-app-2-x-element1--blockClass vtex-app-2-x-element2 vtex-app-2-x-element2--blockClass'
     )
   })
+
   it('should not apply blockClasses if not available', () => {
     const CSS_HANDLES = ['element1', 'element2'] as const
+
     ;(useExtension as any).mockImplementationOnce(() => ({
       component: 'vtex.app@2.1.0',
       props: {},
     }))
 
     const { getByTestId } = renderComponent(CSS_HANDLES)
+
     expect(getByTestId('test-div').className).toBe(
       'vtex-app-2-x-element1 vtex-app-2-x-element2'
     )
   })
+
   describe('migration', () => {
     it('should add both the current app and the migration app', () => {
       const CSS_HANDLES = ['element1', 'element2']
@@ -76,10 +74,12 @@ describe('withCssHandles', () => {
       }
 
       const { getByTestId } = renderComponent(CSS_HANDLES, options)
+
       expect(getByTestId('test-div').className).toBe(
         'vtex-app-2-x-element1 vtex-app-2-x-element1--blockClass vtex-previous-app-3-x-element1 vtex-previous-app-3-x-element1--blockClass vtex-app-2-x-element2 vtex-app-2-x-element2--blockClass vtex-previous-app-3-x-element2 vtex-previous-app-3-x-element2--blockClass'
       )
     })
+
     it('should add more than one migration if needed', () => {
       const CSS_HANDLES = ['element1', 'element2']
       const options = {
@@ -87,6 +87,7 @@ describe('withCssHandles', () => {
       }
 
       const { getByTestId } = renderComponent(CSS_HANDLES, options)
+
       expect(getByTestId('test-div').className).toBe(
         'vtex-app-2-x-element1 vtex-app-2-x-element1--blockClass vtex-previous-app-2-x-element1 vtex-previous-app-2-x-element1--blockClass vtex-previous-app-3-x-element1 vtex-previous-app-3-x-element1--blockClass vtex-app-2-x-element2 vtex-app-2-x-element2--blockClass vtex-previous-app-2-x-element2 vtex-previous-app-2-x-element2--blockClass vtex-previous-app-3-x-element2 vtex-previous-app-3-x-element2--blockClass'
       )
@@ -103,6 +104,7 @@ describe('withCssHandles', () => {
       }
 
       const { getByTestId } = renderComponent(CSS_HANDLES, options)
+
       expect(getByTestId('test-div').className).toBe(
         'vtex-app-2-x-element1 vtex-app-2-x-element1--blockClass vtex-previous-app-2-x-element1 vtex-previous-app-2-x-element1--blockClass vtex-previous-app-3-x-element1 vtex-previous-app-3-x-element1--blockClass vtex-app-2-x-element2 vtex-app-2-x-element2--blockClass vtex-previous-app-2-x-element2 vtex-previous-app-2-x-element2--blockClass vtex-previous-app-3-x-element2 vtex-previous-app-3-x-element2--blockClass'
       )